proyecto_3_chatbot_ai/requirements.txt
nltk
numpy
scikit-learn
google-generativeai